backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.2.0
pandas>=2.0.0
numpy>=1.24.0
pykrx>=0.6.0
requests>=2.31.0
python-dotenv>=1.0.0
